+++ frontend/components_react/event_calendar/event_calendar.jsx
@@ -478,10 +478,10 @@ export default class EventCalendar extends React.Component {
           <div className="buttons">
             <a className="btn btn-primary btn-sm with_right_margin" href={`${this.props.locale === 'en' ? '' : '/' + this.props.locale}/events/${currentEvent.slug}`}>{strings.subscribed}</a>
             {this.props.user_character_ids.includes(currentEvent.owner_id) &&
-              <a className="btn btn-icon btn-edit with_right_margin" href={`${this.props.locale === 'en' ? '' : '/' + this.props.locale}/events/${currentEvent.slug}/edit`}></a>
+              <a className="btn btn-icon btn-edit with_right_margin" href={`${this.props.locale === 'en' ? '' : '/' + this.props.locale}/events/${currentEvent.slug}/edit`} aria-label="Edit button"></a>
             }
             {this.props.user_character_ids.includes(currentEvent.owner_id) &&
-              <button data-confirm={strings.sure} className="btn btn-icon btn-delete" onClick={this._onDeleteEvent.bind(this, currentEvent)}></button>
+              <button data-confirm={strings.sure} className="btn btn-icon btn-delete" onClick={this._onDeleteEvent.bind(this, currentEvent)} aria-label="Delete button"></button>
             }
           </div>
         </div>