+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Modified
 - load wowhead only for need pages
+- aria-label for links without text
 
 ## [1.1.1] - 2019-12-14
 ### Added